+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/GetFollowersCountHandler.java
@@ -1,0 +1,32 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+
+import androidx.annotation.NonNull;
+
+import edu.byu.cs.tweeter.client.model.service.FollowService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.GetFollowersCountTask;
+
+public class GetFollowersCountHandler extends Handler {
+    private FollowService.GetFollowersCountObserver observer;
+
+    public GetFollowersCountHandler(FollowService.GetFollowersCountObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+        boolean success = msg.getData().getBoolean(GetFollowersCountTask.SUCCESS_KEY);
+        if (success) {
+            int count = msg.getData().getInt(GetFollowersCountTask.COUNT_KEY);
+            observer.handleSuccess(count);
+        } else if (msg.getData().containsKey(GetFollowersCountTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(GetFollowersCountTask.MESSAGE_KEY);
+            observer.handleFailure(message);
+        } else if (msg.getData().containsKey(GetFollowersCountTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(GetFollowersCountTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
