+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/IsFollowerHandler.java
@@ -1,0 +1,34 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+
+import androidx.annotation.NonNull;
+
+import edu.byu.cs.tweeter.client.model.service.FollowService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.IsFollowerTask;
+
+public class IsFollowerHandler extends Handler {
+    private FollowService.IsFollowerObserver observer;
+
+    public IsFollowerHandler(FollowService.IsFollowerObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+        boolean success = msg.getData().getBoolean(IsFollowerTask.SUCCESS_KEY);
+        if (success) {
+            boolean isFollower = msg.getData().getBoolean(IsFollowerTask.IS_FOLLOWER_KEY);
+
+            // If logged in user if a follower of the selected user, display the follow button as "following"
+            observer.handleSuccess(isFollower);
+        } else if (msg.getData().containsKey(IsFollowerTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(IsFollowerTask.MESSAGE_KEY);
+            observer.handleFailure(message);
+        } else if (msg.getData().containsKey(IsFollowerTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(IsFollowerTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
