+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/RegisterHandler.java
@@ -1,0 +1,38 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+
+import androidx.annotation.NonNull;
+
+import edu.byu.cs.tweeter.client.model.service.UserService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.RegisterTask;
+import edu.byu.cs.tweeter.model.domain.AuthToken;
+import edu.byu.cs.tweeter.model.domain.User;
+
+public class RegisterHandler extends Handler {
+
+    private UserService.RegisterObserver observer;
+
+    public RegisterHandler(UserService.RegisterObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+
+        boolean success = msg.getData().getBoolean(RegisterTask.SUCCESS_KEY);
+        if (success) {
+            User registeredUser = (User) msg.getData().getSerializable(RegisterTask.USER_KEY);
+            AuthToken authToken = (AuthToken) msg.getData().getSerializable(RegisterTask.AUTH_TOKEN_KEY);
+
+            observer.handleSuccess(registeredUser, authToken);
+        } else if (msg.getData().containsKey(RegisterTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(RegisterTask.MESSAGE_KEY);
+            observer.handleFailure(message);
+        } else if (msg.getData().containsKey(RegisterTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(RegisterTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
