+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/UnfollowHandler.java
@@ -1,0 +1,32 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+
+import androidx.annotation.NonNull;
+
+import edu.byu.cs.tweeter.client.model.service.FollowService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.FollowTask;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.UnfollowTask;
+
+public class UnfollowHandler extends Handler {
+    private FollowService.UnfollowObserver observer;
+
+    public UnfollowHandler(FollowService.UnfollowObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+        boolean success = msg.getData().getBoolean(UnfollowTask.SUCCESS_KEY);
+        if (success) {
+            observer.handleSuccess();
+        } else if (msg.getData().containsKey(FollowTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(FollowTask.MESSAGE_KEY);
+            observer.handleFailure(message);
+        } else if (msg.getData().containsKey(FollowTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(FollowTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
